requirements.txt
# Optional runtime dependencies
spotipy>=2.19.0
Pillow>=8.4.0
colorthief>=0.2.1
requests>=2.28.2
